requirements.txt
duckdb>=1.4.0,<1.5
python-dotenv
dlt[duckdb,motherduck]
kagglehub
kagglehub[pandas-datasets]
dbt-core
dbt-duckdb
pandas